+++ src/custom/global.ts
@@ -1,7 +1,9 @@
+import redis from 'redis'
 import { Firestore } from '@google-cloud/firestore'
 
 interface CustomNodeJSGlobal extends NodeJS.Global {
-  firestoreDB: Firestore,
+  firestoreDB: Firestore
+  redisDB    : redis.RedisClient
 }
 
 export { CustomNodeJSGlobal }