+++ src/database/index.ts
@@ -1,0 +1,4 @@
+import { firebaseConnection } from './firebase'
+import { redisConnection } from './redis'
+
+export { firebaseConnection, redisConnection }